requirements.txt
colorama==0.4.6
dnspython==2.6.1
halo==0.0.31
log-symbols==0.0.14
motor==3.3.2
numpy==1.26.4
pandas==2.2.1
psutil==5.9.8
pymongo==4.6.2
python-dateutil==2.9.0.post0
pytz==2024.1
scapy==2.5.0
six==1.16.0
spinners==0.0.24
termcolor==2.4.0
tornado==6.4
tzdata==2024.1
pyjwt==2.8.0
